+++ task8/CarServiceUI/src/com/senla/carservice/ui/actions/order/CloseOrder.java
@@ -24,7 +24,7 @@ public class CloseOrder implements IAction {
 		Printer.print("Choose Order to cancel");
 
 		List<Object> params = new ArrayList<>();
-		params.add(Reader.readInt());
+		params.add(Reader.readLong());
 		request = new HashMap<>();
 		request.put(Commands.getOrderById, params);
 		responce = ClientHandler.handle(request);
